# docs/swagger.yaml
basePath: /
definitions:
...
  domains.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  domains.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/domains.LoginResponseData'
      message:
        type: string
    type: object
  domains.LoginResponseData:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  domains.RefreshAuthnData:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  domains.RefreshAuthnResponse:
    properties:
      data:
        $ref: '#/definitions/domains.RefreshAuthnData'
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a To-Do List API
  title: To-Do List API
  version: "1.0"
paths:
  /api/auth:
    post:
      consumes:
      - application/json
      description: authenticate user
      parameters:
      - description: email address of the user
        format: email
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: password of the user
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domains.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
      summary: Login with Email & Password
      tags:
      - authentication
    put:
      consumes:
      - application/json
      description: Generating new access token using a refresh token. Only valid refresh
        token will generate new
      parameters:
      - description: refresh token possessed by the user
        in: body
        name: refreshToken
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domains.RefreshAuthnResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domains.ErrorResponse'
      summary: Refresh Authentication
      tags:
      - authentication
...
securityDefinitions:
  "":
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"